separated iex bww module
