+++ src/importexport/CMakeLists.txt
@@ -19,6 +19,7 @@
 
 add_subdirectory(bb)
 add_subdirectory(braille)
+add_subdirectory(bww)
 add_subdirectory(musicxml)
 
 # We are separating the importexport module into separate modules,
@@ -30,7 +31,6 @@ include(${CMAKE_CURRENT_LIST_DIR}/internal/ove/ove.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/internal/guitarpro/guitarpro.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/internal/musedata/musedata.cmake)
 include(${CMAKE_CURRENT_LIST_DIR}/internal/capella/capella.cmake)
-include(${CMAKE_CURRENT_LIST_DIR}/internal/bww/bww.cmake)
 
 set(MODULE_SRC
     ${BRAILLE_SRC}
@@ -39,7 +39,6 @@ set(MODULE_SRC
     ${GUITARPRO_SRC}
     ${MUSEDATA_SRC}
     ${CAPELLA_SRC}
-    ${BWW_SRC}
     ${CMAKE_CURRENT_LIST_DIR}/importexportmodule.cpp
     ${CMAKE_CURRENT_LIST_DIR}/importexportmodule.h
     ${CMAKE_CURRENT_LIST_DIR}/iimportexportconfiguration.h
@@ -54,8 +53,6 @@ set(MODULE_SRC
     ${CMAKE_CURRENT_LIST_DIR}/internal/capellareader.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/overeader.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/overeader.h
-    ${CMAKE_CURRENT_LIST_DIR}/internal/notationbwwreader.cpp
-    ${CMAKE_CURRENT_LIST_DIR}/internal/notationbwwreader.h
     ${CMAKE_CURRENT_LIST_DIR}/internal/guitarproreader.cpp
     ${CMAKE_CURRENT_LIST_DIR}/internal/guitarproreader.h
