+++ src/importexport/bww/CMakeLists.txt
@@ -1,0 +1,37 @@
+#=============================================================================
+#  MuseScore
+#  Music Composition & Notation
+#
+#  Copyright (C) 2020 MuseScore BVBA and others
+#
+#  This program is free software; you can redistribute it and/or modify
+#  it under the terms of the GNU General Public License version 2.
+#
+#  This program is distributed in the hope that it will be useful,
+#  but WITHOUT ANY WARRANTY; without even the implied warranty of
+#  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+#  GNU General Public License for more details.
+#
+#  You should have received a copy of the GNU General Public License
+#  along with this program; if not, write to the Free Software
+#  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
+#=============================================================================
+
+set(MODULE iex_bww)
+
+include(${CMAKE_CURRENT_LIST_DIR}/internal/bww/bww.cmake)
+
+set(MODULE_SRC
+    ${BWW_SRC}
+    ${CMAKE_CURRENT_LIST_DIR}/bwwmodule.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/bwwmodule.h
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notationbwwreader.cpp
+    ${CMAKE_CURRENT_LIST_DIR}/internal/notationbwwreader.h
+    )
+
+set(MODULE_LINK
+    libmscore
+    )
+
+include(${PROJECT_SOURCE_DIR}/build/module.cmake)
+
