+++ src/importexport/bww/internal/bww/symbols.cpp
