+++ src/importexport/bww/internal/bww/symbols.h
