+++ src/importexport/bww/internal/notationbwwreader.cpp
@@ -16,20 +16,19 @@
 //  along with this program; if not, write to the Free Software
 //  Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 //=============================================================================
-#ifndef MU_IMPORTEXPORT_NOTATIONBWWREADER_H
-#define MU_IMPORTEXPORT_NOTATIONBWWREADER_H
+#include "notationbwwreader.h"
 
-#include "notation/inotationreader.h"
+#include "libmscore/score.h"
+#include "notation/notationerrors.h"
 
-namespace mu {
-namespace importexport {
-class NotationBwwReader : public notation::INotationReader
+namespace Ms {
+extern Score::FileError importBww(MasterScore*, const QString& name);
+}
+
+using namespace mu::iex::bww;
+
+mu::Ret NotationBwwReader::read(Ms::MasterScore* score, const io::path& path)
 {
-public:
-
-    Ret read(Ms::MasterScore* score, const io::path& path) override;
-};
+    Ms::Score::FileError err = Ms::importBww(score, path.toQString());
+    return mu::notation::scoreFileErrorToRet(err);
 }
-}
-
-#endif // MU_IMPORTEXPORT_NOTATIONBWWREADER_H
