+++ src/main/CMakeLists.txt
@@ -168,6 +168,7 @@ if (NOT CC_IS_EMSCRIPTEN)
     list(APPEND LINK_LIB iex_musicxml)
     list(APPEND LINK_LIB iex_bb)
     list(APPEND LINK_LIB iex_braille)
+    list(APPEND LINK_LIB iex_bww)
     list(APPEND LINK_LIB plugins)
 endif()
